Fix countof()

+++ Source/WorkerSemaphore.cpp
@@ -32,7 +32,7 @@ namespace Windows {
 
 void WorkerSemaphore::thread_proc (SchedulerSlave& scheduler)
 {
-	while (WaitForMultipleObjects (countof (handles_), handles_, FALSE, INFINITE) == WAIT_OBJECT_0) {
+	while (WaitForMultipleObjects ((DWORD)countof (handles_), handles_, FALSE, INFINITE) == WAIT_OBJECT_0) {
 		scheduler.execute ();
 	}
 }
